kvtransfer: Enable Wall, Werror (#2051)

+++ blade_llm/_C/csrc/kvtransfer/tests/utils_iterator_test.cpp
@@ -7,13 +7,13 @@ using namespace blade_llm;
 TEST(UtilsTest, TestIteratorSource) {
   std::vector<int> nums{1, 2, 3, 4, 5, 6, 7};
   auto iter = Source<int>::from(nums);
-  for (int i = 0; i < nums.size(); ++i) {
+  for (size_t i = 0; i < nums.size(); ++i) {
     auto opt = iter->next();
     EXPECT_TRUE(opt.has_value());
     EXPECT_EQ(*opt.value(), i + 1);
     *opt.value() = 0;
   }
-  for (int i = 0; i < nums.size(); ++i) {
+  for (size_t i = 0; i < nums.size(); ++i) {
     EXPECT_EQ(nums[i], 0);
   }
 }
@@ -22,7 +22,7 @@ TEST(UtilsTest, TestIteratorFilter) {
   std::vector<int> nums{1, 2, 3, 4, 5, 6, 7};
   auto src = Source<int>::from(nums);
   Filter<int *> filter(src, [](int *const &x) { return (*x) % 2 == 0; });
-  for (int i = 0; i < 3; ++i) {
+  for (size_t i = 0; i < 3; ++i) {
     auto opt = filter.next();
     EXPECT_TRUE(opt.has_value());
     EXPECT_EQ(*opt.value(), 2 * i + 2);
@@ -34,7 +34,7 @@ TEST(UtilsTest, TestIteratorMap) {
   std::vector<int> nums{1, 2, 3, 4, 5, 6, 7};
   auto src = Source<int>::from(nums);
   Map<int *, int> map(src, [](int *const &x) { return (*x) * 2; });
-  for (int i = 0; i < nums.size(); ++i) {
+  for (size_t i = 0; i < nums.size(); ++i) {
     auto opt = map.next();
     EXPECT_TRUE(opt.has_value());
     EXPECT_EQ(opt.value(), 2 * i + 2);
@@ -48,7 +48,7 @@ TEST(UtilsTest, TestIteratorChain) {
   auto iter = src
       .filter([](int *const &x) { return *x % 2 == 0; })
       .map<int>([](int *const &x) { return (*x) * 2; });
-  for (int i = 0; i < 3; ++i) {
+  for (size_t i = 0; i < 3; ++i) {
     auto opt = iter->next();
     EXPECT_TRUE(opt.has_value());
     EXPECT_EQ(opt.value(), (2 * i + 2) * 2);
@@ -58,13 +58,13 @@ TEST(UtilsTest, TestIteratorChain) {
 TEST(UtilsTest, TestIteratorItemCopy) {
   std::vector<int> nums{1, 2, 3, 4, 5, 6, 7};
   auto iter = CopySource<int>::from(nums);
-  for (int i = 0; i < nums.size(); ++i) {
+  for (size_t i = 0; i < nums.size(); ++i) {
     auto opt = iter->next();
     EXPECT_TRUE(opt.has_value());
     EXPECT_EQ(opt.value(), i + 1);
     opt.value() = 0;
   }
-  for (int i = 0; i < nums.size(); ++i) {
+  for (size_t i = 0; i < nums.size(); ++i) {
     EXPECT_EQ(nums[i], i + 1);
-  }
+}
 }
